added find_beat_times, returns list of beat times
Two functions which have yet to be written are called in this function:
beat_in_interval: which takes in a time and voltage array, an interval, and an interval end time and returns the time that a beat occured in that interval.
combine_double_beats: if the beat_times array finds two beats close in
time, the higher amplitude time is used for the heartbeat.